requirements.txt
alembic==1.8.0
blinker==1.4
Cerberus==1.3.4
click==8.1.3
Flask==2.1.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
